requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.3.3
python-dotenv==1.0.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6